requirements.txt
# ============================================================================
# OCRS Backend - Python Dependencies
# ============================================================================

# Web Framework
Flask==3.0.0
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0

# Database
mysql-connector-python==8.2.0
PyMySQL==1.1.0

# Password Hashing
bcrypt==4.1.2

# Environment Variables
python-dotenv==1.0.0

# Input Validation
email-validator==2.1.0
python-multipart==0.0.6

# Date/Time Utilities
python-dateutil==2.8.2

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0

# Code Quality
pylint==3.0.3
black==23.12.1
flake8==7.0.0

# API Documentation
flasgger==0.9.7.1

# Logging
python-json-logger==2.0.7

# Rate Limiting
Flask-Limiter==3.5.0

# Utilities
requests==2.31.0